The great separation of concerns rewrite...
