+++ gwf/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, print_function
 
-from gwf.core import Event, PreparedWorkflow, Target, Workflow
+from gwf.core import PreparedWorkflow, Target, Workflow
 
 USER_CONFIG_FILE = '~/.gwfrc'
 LOCAL_CONFIG_FILE = '.gwfrc'
@@ -34,4 +34,4 @@ class template(object):
             self.__class__.__name__, self.options, self.spec)
 
 
-__all__ = ('Target', 'Workflow', 'PreparedWorkflow', 'template', 'Event',)
+__all__ = ('Target', 'Workflow', 'PreparedWorkflow', 'template',)
