+++ tests/backends/test_slurm.py
@@ -54,19 +54,19 @@ class TestSlurmBackendConfigure(SlurmTestCase):
 
     def test_does_not_create_log_dir_if_it_already_exists(self):
         self.mock_exists.return_value = True
-        self.backend.configure(workflow=self.prepared_workflow,
+        self.backend.configure(working_dir='/some/dir',
                                config=self.config)
         self.mock_makedirs.assert_not_called()
 
     def test_creates_log_dir_if_it_does_not_already_exist(self):
         self.mock_exists.return_value = False
-        self.backend.configure(workflow=self.prepared_workflow,
+        self.backend.configure(working_dir='/some/dir',
                                config=self.config)
         self.mock_makedirs.assert_called_once_with('/some/dir/.gwf/logs')
 
     def test_jobdb_is_empty_if_job_db_file_does_not_exist(self):
         self.mock_exists.return_value = True
-        self.backend.configure(workflow=self.prepared_workflow,
+        self.backend.configure(working_dir='/some/dir',
                                config=self.config)
         self.mock_read_json.assert_any_call(
             '.gwf/slurm-backend-jobdb.json')
@@ -75,7 +75,7 @@ class TestSlurmBackendConfigure(SlurmTestCase):
     def test_jobdb_is_loaded_from_job_db_file_when_it_exists(self):
         self.mock_read_json.return_value = {"TestTarget": '1000'}
         self.mock_get_live_job_states.return_value = {'1000': 'R'}
-        self.backend.configure(workflow=self.prepared_workflow,
+        self.backend.configure(working_dir='/some/dir',
                                config=self.config)
         self.mock_read_json.assert_any_call('.gwf/slurm-backend-jobdb.json')
         self.assertDictEqual(self.backend._job_db, {'TestTarget': '1000'})
@@ -136,8 +136,9 @@ class TestSlurmBackendSubmit(SlurmTestCase):
             config={},
         )
 
-        backend.configure(workflow=prepared_workflow, config=self.config)
-        backend.submit(target3)
+        backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
+        backend.submit(target3, prepared_workflow.dependencies[target3])
 
         self.mock_call_sbatch.assert_any_call(ANY, ['1000', '2000'])
         self.assertEqual(backend._job_db['TestTarget3'], '3000')
@@ -157,8 +158,9 @@ class TestSlurmBackendSubmit(SlurmTestCase):
         )
 
         backend = SlurmBackend()
-        backend.configure(workflow=prepared_workflow, config=self.config)
-        backend.submit(target)
+        backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
+        backend.submit(target, prepared_workflow.dependencies[target])
 
         self.mock_call_sbatch.assert_any_call(ANY, [])
         self.assertEqual(backend._job_db['TestTarget'], '1000')
@@ -173,7 +175,8 @@ class TestSlurmBackendSubmit(SlurmTestCase):
         )
 
         backend = SlurmBackend()
-        backend.configure(workflow=prepared_workflow, config=self.config)
+        backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         target = Target(
             name='TestTarget',
@@ -231,7 +234,8 @@ class TestSlurmBackendCancel(SlurmTestCase):
         )
 
         backend = SlurmBackend()
-        backend.configure(workflow=prepared_workflow, config=self.config)
+        backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
         backend.cancel(target)
 
         self.mock_call_scancel.assert_any_call('1000')
@@ -252,7 +256,8 @@ class TestSlurmBackendCancel(SlurmTestCase):
         )
 
         backend = SlurmBackend()
-        backend.configure(workflow=prepared_workflow, config=self.config)
+        backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         with self.assertRaises(BackendError):
             backend.cancel(target)
@@ -276,7 +281,8 @@ class TestSlurmBackendSubmitted(SlurmTestCase):
             supported_options=SlurmBackend.supported_options,
             config={},
         )
-        self.backend.configure(workflow=prepared_workflow, config=self.config)
+        self.backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         self.assertTrue(self.backend.submitted(target1))
         self.assertFalse(self.backend.submitted(target2))
@@ -301,7 +307,8 @@ class TestSlurmBackendRunning(SlurmTestCase):
             supported_options=SlurmBackend.supported_options,
             config={},
         )
-        self.backend.configure(workflow=prepared_workflow, config=self.config)
+        self.backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         self.assertTrue(self.backend.running(target1))
         self.assertFalse(self.backend.running(target2))
@@ -318,7 +325,8 @@ class TestSlurmBackendLogs(SlurmTestCase):
             supported_options=SlurmBackend.supported_options,
             config={},
         )
-        self.backend.configure(workflow=prepared_workflow, config=self.config)
+        self.backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         with self.assertRaises(NoLogFoundError):
             self.backend.logs(target)
@@ -338,7 +346,8 @@ class TestSlurmBackendLogs(SlurmTestCase):
             supported_options=SlurmBackend.supported_options,
             config={},
         )
-        self.backend.configure(workflow=prepared_workflow, config=self.config)
+        self.backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         m = mock_open(read_data='this is the log file')
         with patch('builtins.open', m):
@@ -362,7 +371,8 @@ class TestSlurmBackendLogs(SlurmTestCase):
             supported_options=SlurmBackend.supported_options,
             config={},
         )
-        self.backend.configure(workflow=prepared_workflow, config=self.config)
+        self.backend.configure(
+            working_dir=prepared_workflow.working_dir, config=self.config)
 
         m = mock_open()
         m.side_effect = [
@@ -381,7 +391,7 @@ class TestSlurmBackendClose(SlurmTestCase):
 
     def test_closing_backend_dumps_database_atomically(self):
         self.backend.configure(
-            workflow=self.prepared_workflow,
+            working_dir='/some/dir',
             config=self.config
         )
         self.backend.close()
